tools/cellxgene_census_builder/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cellxgene_census_builder"
dynamic = ["version"]
description = "Build Cell Census"
authors = [
    { name = "Chan Zuckerberg Initiative Foundation", email = "soma@chanzuckerberg.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = "~= 3.11"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.11",
]
dependencies= [
    "typing_extensions==4.9.0",
    "pyarrow==14.0.2",
    "pandas[performance]==2.1.4",
    "anndata==0.10.3",
    "numpy==1.23.5",
    # IMPORTANT: consider TileDB format compat before advancing this version. It is important that
    # IMPORTANT: the tiledbsoma version lag that used in cellxgene-census package.
    "tiledbsoma==1.6.1",
    "cellxgene-census==1.9.1",
    "scipy==1.11.4",
    "fsspec==2023.12.2",
    "s3fs==2023.12.2",
    "requests==2.31.0",
    "aiohttp==3.9.1",
    "Cython", # required by owlready2
    "wheel",  # required by owlready2
    "owlready2==0.44",
    "gitpython==3.1.41",
    "attrs==23.1.0",
    "psutil==5.9.6",
    "pyyaml==6.0.1",
    "numba==0.57",
]

[project.urls]
homepage = "https://github.com/chanzuckerberg/cellxgene-census"
repository = "https://github.com/chanzuckerberg/cellxgene-census"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cellxgene_census_builder*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)

[tool.setuptools_scm]
root = "../.."

[tool.black]
line-length = 120
target_version = ['py311']

[tool.mypy]
show_error_codes = true
ignore_missing_imports = true
warn_unreachable = true
strict = true
plugins = "numpy.typing.mypy_plugin"

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = ["E501", "E402", "C408", ]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
markers = [
    "live_corpus: runs on the live Census data corpus and small enough to run in CI",
    "expensive: too expensive to run regularly or in CI",
]
